+++ client/src/components/card/Filter.tsx
@@ -7,19 +7,20 @@ import axios from "@/utils/axios";
 import { toast } from "react-toastify";
 
 type Props = {
-    setFilterData: any
+  setFilterData: any;
 };
 
-const Filter = ({setFilterData}: Props) => {
+const Filter = ({ setFilterData }: Props) => {
   const [hotels, setHotels] = React.useState<any>([]);
-  const [filter,setFilter] = useState({
+  const [filter, setFilter] = useState({
     guestName: "",
-    hotelName: "",
-    bookingSource: "",
+    hotelName: "--select--",
+    bookingSource: "--select--",
     serialNumber: "",
-    filterBy: "",
-    dateRange: {}
-  })
+    filterBy: "--select--",
+    status: "--select--",
+    dateRange: {},
+  });
   const [isFilterOptionSelected, setisFilterOptionSelected] =
     useState<boolean>(false);
   const [loading, setLoading] = React.useState<boolean>(false);
@@ -54,16 +55,13 @@ const Filter = ({setFilterData}: Props) => {
   const handleSelect = (ranges: any) => {
     // console.log(ranges)
     setSelectionRange(ranges.selection);
-    filter.dateRange = ranges.selection
-    
+    filter.dateRange = ranges.selection;
   };
-
 
   const handleSubmit = () => {
     // console.log(filter)
-    setFilterData(filter)
-
-  }
+    setFilterData(filter);
+  };
   return (
     <div className="rounded-md shadow-lg w-full flex justify-start items-start p-4 flex-col">
       <h1 className="font-bold text-xl mb-2">Filter</h1>
@@ -76,7 +74,10 @@ const Filter = ({setFilterData}: Props) => {
                 Guest name
               </label>
               <input
-              onChange={(e) => filter.guestName = e.target.value}
+              value={filter.guestName}
+                onChange={(e) => {
+                  setFilter({...filter, guestName: e.target.value})
+                }}
                 placeholder="Enter guest name"
                 type="text"
                 className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
@@ -87,11 +88,17 @@ const Filter = ({setFilterData}: Props) => {
                 Hotel name
               </label>
               <select
-              onChange={(e) => filter.hotelName = e.target.value}
+                onChange={(e) => {
+                  setFilter({...filter, hotelName: e.target.value})
+                }}
                 name=""
-                id=""
+                id="hotel-drop-down"
                 className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+                value={filter.hotelName && filter.hotelName}
               >
+                <option disabled selected value={"--select--"}>
+                  --Select--
+                </option>
                 {hotels.map((hotel: any, index: any) => {
                   return (
                     <option key={hotel._id} value={hotel._id}>
@@ -106,12 +113,18 @@ const Filter = ({setFilterData}: Props) => {
                 Booking source
               </label>
               <select
-                onChange={(e) => filter.bookingSource= e.target.value}
-                id="paymentby"
-                name="paymentby"
+                onChange={(e) => {
+                  setFilter({...filter, bookingSource: e.target.value})
+                }}
+                id="source-drop-down"
+                name=""
+
                 className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+                value={filter.bookingSource}
               >
-                <option defaultValue="choose">Choose</option>
+                <option disabled selected value={'--select--'}>
+                  --Select--
+                </option>
                 <option value="Booking.com">Booking.com</option>
                 <option value="Agoda">Agoda</option>
                 <option value="Cleartrip">Cleartrip</option>
@@ -131,29 +144,60 @@ const Filter = ({setFilterData}: Props) => {
                 Sl. no
               </label>
               <input
-              onChange={(e) => filter.serialNumber = e.target.value}
+                onChange={(e) => {
+                  setFilter({...filter, serialNumber: e.target.value})
+                }}
                 type="number"
+                value={filter.serialNumber}
+                placeholder="Enter serial number"
                 className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
               />
             </div>
             <div className="">
               {" "}
-              <h1>Select</h1>
+              <h1>Status</h1>
               <select
-              onChange={(e) => {setisFilterOptionSelected(true); filter.filterBy = e.target.value} }
+                onChange={(e) => {
+                  setisFilterOptionSelected(true);
+                  setFilter({...filter, status: e.target.value})
+                }}
                 name=""
-                id=""
                 className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+                value={filter.status && filter.status}
+
               >
-                <option defaultValue="select">
+                <option disabled selected value={'--select--'}>
+                  --Select--
+                </option>
+                <option value="CONFIRMED">Confirmed</option>
+                <option value="CANCELLED">Cancelled</option>
+
+
+              </select>
+            </div>
+            <div className="">
+              {" "}
+              <h1>Date Range By</h1>
+              <select
+                onChange={(e) => {
+                  setisFilterOptionSelected(true);
+                  setFilter({...filter, filterBy: e.target.value})
+                }}
+                name=""
+                id="filterBy-drop-down"
+                className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+                value={filter.filterBy && filter.filterBy}
+
+              >
+                <option disabled selected value={'--select--'}>
                   --Select--
                 </option>
                 <option value="createdAt">Created</option>
-                <option value="checkIndate">Arrival</option>
+                <option value="checkInDate">Arrival</option>
 
                 <option value="checkOutDate">Departure</option>
-
                 <option value="updatedAt">Cancelled</option>
+
               </select>
             </div>
           </div>
@@ -169,13 +213,43 @@ const Filter = ({setFilterData}: Props) => {
           </div>
         )}
       </div>
-      <div className="mt-2">
+      <div className="mt-2 flex gap-2">
         <button
-        onClick={handleSubmit}
+          onClick={handleSubmit}
           type="submit"
+          disabled={
+            !filter.guestName && filter.hotelName === "--select--" && filter.bookingSource === "--select--" && !filter.serialNumber && filter.filterBy === "--select--" && filter.status === "--select--"
+          }
           className=" text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm w-full sm:w-auto px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800 disabled:opacity-50"
         >
           Search
+        </button>
+        <button
+          onClick={()=>{
+            setisFilterOptionSelected(false)
+            setFilter({
+              guestName: "",
+              hotelName: "--select--",
+              bookingSource: "--select--",
+              serialNumber: "",
+              filterBy: "--select--",
+              dateRange: {},
+              status: "--select--",
+            })
+            setFilterData({
+              guestName: "",
+              hotelName: "",
+              bookingSource: "",
+              serialNumber: "",
+              filterBy: "",
+              dateRange: {},
+              status: "",
+            })
+          }}
+          type="submit"
+          className=" text-blue-700 border-2 border-blue-700 inset-2 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm w-full sm:w-auto px-5 py-2.5 text-center dark:bg-blue-600 dark:text-white dark:hover:bg-blue-700 disabled:opacity-50 hover:text-white"
+        >
+          Reset
         </button>
       </div>
     </div>